apps/perception-photodna/requirements.txt
pillow==12.1.1
uvicorn==0.41.0
fastapi==0.135.1